chore: move postgres_command.go to command module
